Almost finishing
In this version tree.py was addict,

Tree.py transform the data that build the maze in a tree and find the path to exit.

Readme was created with informations about this project

Because of tree.py, most of the code in mouse.py became useless, but I prefer to keep it to show my professor my errors too.

Finally some alterations about speed and esthetics.